// pos/src/app/models/m-company-master.ts
export class MCompanyMaster {
    constructor() { }
    
    id?: number;        
    companyCode?: string;        
    companyName?: string;        
    address?: string;        
    countrySettingID?: number;        
    retailSettingID?: number;
    remarks?: string;          
    companyLogo?: string;          
    countryName?: string;          
    retailName?: string;          
    retailSettingCode?: string;          
    countrySettingCode?: string;
    active?:Boolean;
}